requirements.txt
huggingface_hub
ultralytics==8.0.124
timm==0.8.13.dev0
yt_dlp
lapx>=0.5.2
requests
Django==3.2.12
djangorestframework==3.12.4
pyrebase4==4.5.0
Pillow==8.4.0
torch==1.10.0
firebase-admin==5.3.0
PyJWT==2.8.0
drf-extra-fields==3.7.0
omegaconf==2.3.0